--- src/main/java/jpaEx/config/QueryDSLConfiguration.java
@@ -1,21 +1,0 @@
-package jpaEx.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-//JpaQueryFactory Bean 주입용 클래스
-@Configuration
-public class QueryDSLConfiguration {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
