--- src/main/java/jpaEx/domain/diary/diet/EatTime.java
@@ -1,6 +1,0 @@
-package jpaEx.domain.diary.diet;
-
-public enum EatTime {
-
-    BreakFast, Lunch, Dinner, Else
-}
