--- src/main/java/jpaEx/domain/diary/writer/Role.java
@@ -1,5 +1,0 @@
-package jpaEx.domain.diary.writer;
-
-public enum Role {
-    User
-}
