--- src/main/java/jpaEx/domain/diary/writer/WriterRepositoryCustom.java
@@ -1,9 +1,0 @@
-package jpaEx.domain.diary.writer;
-
-import org.springframework.data.jpa.repository.Query;
-
-public interface WriterRepositoryCustom {
-    Long findCountOfId();
-
-    Long findMaxOfId();
-}
