+++ src/main/java/refactoringAPI/config/QueryDSLConfiguration.java
@@ -1,0 +1,21 @@
+package refactoringAPI.config;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+//JpaQueryFactory Bean 주입용 클래스
+@Configuration
+public class QueryDSLConfiguration {
+
+    @PersistenceContext
+    private EntityManager entityManager;
+
+    @Bean
+    public JPAQueryFactory jpaQueryFactory() {
+        return new JPAQueryFactory(entityManager);
+    }
+}
