+++ src/main/java/refactoringAPI/domain/diary/diet/EatTime.java
@@ -1,0 +1,6 @@
+package refactoringAPI.domain.diary.diet;
+
+public enum EatTime {
+
+    BreakFast, Lunch, Dinner, Else
+}
