+++ src/main/java/refactoringAPI/domain/diary/writer/Role.java
@@ -1,0 +1,5 @@
+package refactoringAPI.domain.diary.writer;
+
+public enum Role {
+    User
+}
