+++ src/main/java/refactoringAPI/domain/diary/writer/WriterRepositoryCustom.java
@@ -1,0 +1,7 @@
+package refactoringAPI.domain.diary.writer;
+
+public interface WriterRepositoryCustom {
+    Long findCountOfId();
+
+    Long findMaxOfId();
+}
